pyproject.toml
[project]
name = "qso"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
description = "Quantum Stochastic Optimization"
version = "0.1.0"
authors = [
    { name = "Arvin Kushwaha", email = "me@arvinsk.org" }
]
dependencies = [
    "pennylane",
    "numpy",
    "jax",
    "jaxlib",
    "orjson",
    "pennylane-lightning",
    "matplotlib",
    "matplotlib-backend-sixel",
    "types-tqdm",
    "pyserde",
]

[project.optional-dependencies]
dev = [
    "tox",
    "pre-commit",
    "bump2version",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["data"]

[[tool.mypy.overrides]]
module = "pennylane.*"
ignore_missing_imports = true
